:recycle: PR Code Refactor
