+++ main/src/main/resources/wsa.beta.properties
@@ -22,3 +22,5 @@ flag-post-guide-id=1072845227418714193
 water-ball-journal-post-id=1072869148826292234
 water-ball-lose-weight-post-id=1091190313575526400
 wsa-guideline-channel-id=1042774419371720715
+wsa-long-article-role-id=1163842900933742592
+wsa-topic-master-role-id=1163842370018754560
